+++ app/constants/constants.js
@@ -2,12 +2,6 @@
 /* eslint-disable no-multi-str */
 import {concat} from 'ramda';
 
-/**
- * Default to '1 week' for `refreshInterval` for backwards compatability. For
- * older versions this will prevent issues such as `setInterval(runQuery, NaN)`
- */
-export const DEFAULT_REFRESH_INTERVAL = 7 * 24 * 60 * 60;
-
 export const DIALECTS = {
     MYSQL: 'mysql',
     MARIADB: 'mariadb',